schema-check: stop printing to stderr
We are abusing stderr, in that we really should just use it for
unexpected errors. Since this is a schema checker, reporting that a
schema is invalid is expected behavior, so it should go to stdout.

Philosophy aside, if you print to stderr, it makes piping to less and
similar programs very annoying, as you have to also redirect stderr.