schema: expose both raw and SI units
Currently, we use only raw units, but it would be useful to standardize
on SI units. Do so, guaranteeing to the user that the raw value rendered
will be an SI unit but both units will be advertised.

Since we will be converting units, we need to express values as floats
because the conversions will not produce clean numbers. Change the
public interface to stop advertising types and number of interpreted
bytes, as the result will always be a float in SI units.