add mode and pid to foreach callback
Sometimes a caller may wish to know the mode and PID for the descriptor
passed into the callback. We could add this information to the
descriptor itself, but it's redundant for most of the other usage of the
API, in which you request a descriptor for a given mode and PID, so you
already know the mode and PID for the descriptor.