parser: switch to document-based parsing
When I first wrote the parser, I didn't realize libYAML could do
document-based (instead of event-based) parsing. So I reluctantly wrote
a state machine, looked at the code, and felt sad. It worked and was
fast, but it was hard to maintain and read.

I have since learned that we can do document-based parsing, which is
much more natural, so let's switch to it.

While we're at it, we can replace the unit strings with unit enums and
substantially simplify the unit parsing logic (which was more complex
than it should be largely because of how the state machine worked).