stack: more stack evaluation thread-safe
Currently, we don't lock the yobd ctx because everything is read-only
reentrant. The one exception here is that each context contains a stack,
so we can't do multiple PID evaluations safely at the some time.

Fix this by putting the data for the evaluation stack on the stack
(haha). Now PID evaluation will be reentrant and still need to no lock.