units: reconcile schema and parser units
We were including all sschema-supported units, but we didn't have
conversion functions for all of them, which can cause runtime assertion
errors. Stop depending on sschema for units and explicitly list the
supported units. In this way, when a schema fails because we introduce a
new unit, we'll know to add a conversion function at that point.