types: reconcile schema and parser types
We were including all sschema-supported types, but we don't actually
support all of them.  which can cause runtime assertion errors. Stop
depending on sschema for types and explicitly list the supported types.

Also, sort the list of types so that we can easily compare against the
schema and make sure we didn't miss anything.