expr: fix precedence bug in shunting yard
Currently, we're hitting a bug in the shunting yard expression parsing
when we see an expression like "A / 2 / 2". Specifically, it's parsing
it as "A / (2/1)", or "A / 1", or just "A". The problem is that the
shunting yard algorithm calls for popping operators from the stack when
an operator is parsed and the top of the operator stack has the same
precedence as the parsed operator. Instead of implementing precedence,
we were parsing just doing alternate switch statements for */ vs +-.
This breaks when we encounter // or ** (or ++ or -- technically).

Fix this by treating */+- the same way and properly implementing
precedence.