parser: fix schema absolute path check
Sometimes, we may wish to specify a schema in a subdirectory of the
schema dir, such as "toyota/schema.yaml" instead of just "schema.yaml"
or a full path. Thus we really should just check if the schema is an
absolute vs relative path instead of checking for "/" anywhere in the
string.